refactor(api): transform complementary certif scoring event handler to usecase
